code refactor: separate pkg for sstable + bugfix + test update for offset length
